+++ ksa_compliance/jinja.py
@@ -6,10 +6,10 @@ from typing import Optional, cast
 
 import frappe
 import pyqrcode
+from erpnext.accounts.doctype.journal_entry.journal_entry import JournalEntry
 from erpnext.accounts.doctype.payment_entry.payment_entry import PaymentEntry
 from erpnext.accounts.doctype.pos_invoice.pos_invoice import POSInvoice
 from erpnext.accounts.doctype.sales_invoice.sales_invoice import SalesInvoice
-from erpnext.accounts.doctype.journal_entry.journal_entry import JournalEntry
 from erpnext.setup.doctype.branch.branch import Branch
 from frappe.utils import flt
 from frappe.utils.data import get_time, getdate
@@ -158,14 +158,14 @@ def get_phase_2_print_format_details(
         customer = payment_entry.party
         advance_payment_entry = get_advance_payment_entry_info(payment_entry, settings)
         customer_id = payment_entry.party
-        net_amount =calculate_net_from_gross_included_in_print_rate(
+        net_amount = calculate_net_from_gross_included_in_print_rate(
             sales_invoice.accounts[0].debit_in_account_currency,
             advance_payment_entry.tax_rate,
         )
         tax_amount = calculate_tax_amount_included_in_print_rate(
             sales_invoice.accounts[0].debit_in_account_currency,
             net_amount,
-        )    
+        )
     else:
         customer = sales_invoice.customer
         customer_id = getattr(sales_invoice, "customer", None)